feature: e2e tests describing how software currently works
There are lots of things to consider here. Business rules are not being enforced properly, example: User is able to change the order status by performing 2 different actions (update and complete endpoints) or you can delete an order at any given time. There is no way to enforce standarization, example: Using a constant/enum to list the accepted order statuses or a list with accepted discount codes. With that being said: app is not applying what are considered as good software guidelines such as SRP or expressing behavior