feature: move logic AWAY from controllers
Was about to refactor the whole app, but decided against it for 2 reasons: 1. It would defeat the purpose of the exercise, this app should be treated as a productive software that's making money and supporting a business, can't just blindy refactor. 2. Complexity, it's easier to refactor by small increments. I'm taking small steps and using the tests as a safety method to keep refactoring.