feature: business logic on aggregate
Now the business logic belongs to an aggregate, in this way I'll be able to enforce business rules on all orders. Next steps are to enforce granularity on aggregate methods, nowadays there is a CRUD like approach to dealing with an order, which is not bad, but it's much better and business like to express behavior, instead of a whole update method, make it smaller: complete-order, apply-discount, change-shipping-address, so on and so forth. Couple of things to consider: -At the moment, I'm not using IoC and DI to their full extent on purpose, too much set up, I wanna explore my solution first. -There is a need of domain services for certain actions, for example: finding orders by id and throwing and error if there is no order.